Added "create tables", query, and index file for Project
The "create tables" file currently includes all the discussed tables. It creates the database and makes all the empty tables. I did this for my own convenience. The "SQL Statements" file contains a list of queries that may be relevant, focusing on the Project relation. The "Index" file displays all the indexes for the Project relation based on our current schema and what attributes I think will be frequently accessed.